--- rlkit/envs/wrappers/proxy_env.py
@@ -1,3 +1,0 @@
-
-
-
