+++ rlkit/samplers/util.py
@@ -1,9 +1,4 @@
 import numpy as np
-from rlkit.samplers.rollout_functions import deprecated_rollout as normal_rollout
-
-def rollout(*args, **kwargs):
-    # TODO Steven: remove pointer
-    return normal_rollout(*args, **kwargs)
 
 def split_paths(paths):
     """
