--- rlkit/torch/networks/deprecated_feedforward.py
@@ -1,104 +1,0 @@
-import torch
-from torch import nn as nn
-from torch.nn import functional as F
-
-from rlkit.pythonplusplus import identity
-from rlkit.torch import pytorch_util as ptu
-from rlkit.torch.core import PyTorchModule, eval_np
-
-
-class FeedForwardQFunction(PyTorchModule):
-    def __init__(
-            self,
-            obs_dim,
-            action_dim,
-            observation_hidden_size,
-            embedded_hidden_size,
-            init_w=3e-3,
-            output_activation=identity,
-            hidden_init=ptu.fanin_init,
-            batchnorm_obs=False,
-    ):
-        print("WARNING: This class will soon be deprecated.")
-        super().__init__()
-
-        self.obs_dim = obs_dim
-        self.action_dim = action_dim
-        self.observation_hidden_size = observation_hidden_size
-        self.embedded_hidden_size = embedded_hidden_size
-        self.hidden_init = hidden_init
-        self.obs_fc = nn.Linear(obs_dim, observation_hidden_size)
-        self.embedded_fc = nn.Linear(observation_hidden_size + action_dim,
-                                     embedded_hidden_size)
-
-        self.last_fc = nn.Linear(embedded_hidden_size, 1)
-        self.output_activation = output_activation
-
-        self.init_weights(init_w)
-        self.batchnorm_obs = batchnorm_obs
-        if self.batchnorm_obs:
-            self.bn_obs = nn.BatchNorm1d(obs_dim)
-
-    def init_weights(self, init_w):
-        self.hidden_init(self.obs_fc.weight)
-        self.obs_fc.bias.data.fill_(0)
-        self.hidden_init(self.embedded_fc.weight)
-        self.embedded_fc.bias.data.fill_(0)
-        self.last_fc.weight.data.uniform_(-init_w, init_w)
-        self.last_fc.bias.data.uniform_(-init_w, init_w)
-
-    def forward(self, obs, action):
-        if self.batchnorm_obs:
-            obs = self.bn_obs(obs)
-        h = obs
-        h = F.relu(self.obs_fc(h))
-        h = torch.cat((h, action), dim=1)
-        h = F.relu(self.embedded_fc(h))
-        return self.output_activation(self.last_fc(h))
-
-
-class FeedForwardPolicy(PyTorchModule):
-    def __init__(
-            self,
-            obs_dim,
-            action_dim,
-            fc1_size,
-            fc2_size,
-            init_w=1e-3,
-            hidden_init=ptu.fanin_init,
-    ):
-        print("WARNING: This class will soon be deprecated.")
-        super().__init__()
-
-        self.obs_dim = obs_dim
-        self.action_dim = action_dim
-        self.fc1_size = fc1_size
-        self.fc2_size = fc2_size
-        self.hidden_init = hidden_init
-
-        self.fc1 = nn.Linear(obs_dim, fc1_size)
-        self.fc2 = nn.Linear(fc1_size, fc2_size)
-        self.last_fc = nn.Linear(fc2_size, action_dim)
-
-        self.init_weights(init_w)
-
-    def init_weights(self, init_w):
-        self.hidden_init(self.fc1.weight)
-        self.fc1.bias.data.fill_(0)
-        self.hidden_init(self.fc2.weight)
-        self.fc2.bias.data.fill_(0)
-
-        self.last_fc.weight.data.uniform_(-init_w, init_w)
-        self.last_fc.bias.data.uniform_(-init_w, init_w)
-
-    def forward(self, obs):
-        h = F.relu(self.fc1(obs))
-        h = F.relu(self.fc2(h))
-        return F.tanh(self.last_fc(h))
-
-    def get_action(self, obs_np):
-        actions = self.get_actions(obs_np[None])
-        return actions[0, :], {}
-
-    def get_actions(self, obs):
-        return eval_np(self, obs)
